requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
spacy
python-dateutil
email-validator
phonenumbers